fix export and rename it's options
