+++ src/main.rs
@@ -6,6 +6,7 @@ mod context;
 mod date_time;
 mod duration;
 mod error;
+mod export;
 mod job;
 mod job_list;
 mod jobs;
